# src/main/resources/openapi.yml
openapi: 3.1.0
info:
  title: Simple Task Scheduler - OpenAPI 3.1
  description: |-
    This is a Task Scheduler application where we have some basic functionalities like 
    - Creating a new Task
    - Creating a new User
    - Assigning a task to a user
    - Changing the status of a task
    - Changing the progress percentage of a task

  termsOfService: NONE
  contact:
    email: yadavvi91@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
paths:
  /api/v1/signup:
    # NO NUMBER
    # Anyone can hit this api
    post:
      summary: Create a Spring Security User
      description: Creates a new Spring Security User and a corresponding Task User
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              ref: "#/components/schemas/SignUpRequest"
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid User Details provided

  # /api/v1/token:
  /api/v1/login:
    # NO NUMBER
    # Anyone can hit this api
    post:
      summary: Generate JWT Token
      description: Generates a JWT token for authentication
      operationId: generateJwtToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JwtTokenRequest"
      responses:
        '200':
          description: JWT Token generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtTokenResponse"
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    SignUpRequest:
      required:
        - username
        - email
        - password
        - firstName
        - lastName
      type: object
      properties:
        username:
          maxLength: 20
          minLength: 3
          type: string
        email:
          maxLength: 50
          minLength: 10
          type: string
        password:
          maxLength: 40
          minLength: 6
          type: string
        firstName:
          maxLength: 20
          minLength: 3
          type: string
        lastName:
          maxLength: 20
          minLength: 3
          type: string
...
    JwtTokenRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    JwtTokenResponse:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
...